updates: embassy rp 0.2.0 et al
Bump the embassy to the package versions from the new examples.
`make_static!` macro is broken these days because of upstream rust
changes, fix was to tediously switch to `StaticCell` which polutes the
main with type information, ugly but necessary.

Update fixes this annoyance bug
https://github.com/embassy-rs/embassy/issues/1736. Preliminary tests on
device seem to be good.